# packages/symbol-rest/specification/swagger.yml
openapi: 3.0.2
info:
  version: 1.0.3
  title: Catapult REST Endpoints
  description: OpenAPI Specification of catapult-rest
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
...
paths:
...
  "/accounts/{accountId}":
    get:
      tags:
        - Account routes
      summary: Get account information
      description: Returns the account information.
      operationId: getAccountInfo
      parameters:
        - $ref: "#/components/parameters/accountIdPath"
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountInfoDTO"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
        "409":
          $ref: "#/components/responses/InvalidArgument"
...
  /transactions:
    put:
      tags:
        - Transaction routes
      summary: Announce a new transaction
      description: |
        Announces a transaction to the network.
        The [catbuffer library](https://github.com/nemtech/catbuffer) defines the protocol to serialize and deserialize Symbol entities.
        Catbuffers are integrated into [Symbol SDKs](https://nemtech.github.io/sdk.html). 
        It's recommended to use SDKs instead of calling the API endpoint directly to announce transactions.
      operationId: announceTransaction
      requestBody:
        $ref: "#/components/requestBodies/transactionPayload"
      responses:
        "202":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnounceTransactionInfoDTO"
        "400":
          $ref: "#/components/responses/InvalidContent"
        "409":
          $ref: "#/components/responses/InvalidArgument"
  /transactions/confirmed:
    get:
      tags:
        - Transaction routes
      summary: Search confirmed transactions
      description: |
        Returns an array of confirmed transactions.
        If a transaction was announced with an alias rather than an address, the address that will be considered when querying is the one that
        was resolved from the alias at confirmation time.
      operationId: searchConfirmedTransactions
      parameters:
        - $ref: "#/components/parameters/addressQuery"
        - $ref: "#/components/parameters/recipientAddressQuery"
        - $ref: "#/components/parameters/signerPublicKeyQuery"
        - $ref: "#/components/parameters/heightQuery"
        - $ref: "#/components/parameters/fromHeightQuery"
        - $ref: "#/components/parameters/toHeightQuery"
        - $ref: "#/components/parameters/fromTransferAmountQuery"
        - $ref: "#/components/parameters/toTransferAmountQuery"
        - $ref: "#/components/parameters/transactionTypeQuery"
        - $ref: "#/components/parameters/embedded"
        - $ref: "#/components/parameters/transferMosaicIdQuery"
        - $ref: "#/components/parameters/pageSizeQuery"
        - $ref: "#/components/parameters/pageNumberQuery"
        - $ref: "#/components/parameters/offsetQuery"
        - $ref: "#/components/parameters/orderQuery"
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionPage"
        "409":
          $ref: "#/components/responses/InvalidArgument"
    post:
      tags:
        - Transaction routes
      summary: Get confirmed trasactions information
      description: Returns confirmed transactions information for a given array of transactionIds.
      operationId: getConfirmedTransactions
      requestBody:
        $ref: "#/components/requestBodies/transactionIds"
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                type: array
                description: Array of transactions information.
                items:
                  $ref: "#/components/schemas/TransactionInfoDTO"
        "400":
          $ref: "#/components/responses/InvalidContent"
        "409":
          $ref: "#/components/responses/InvalidArgument"
  "/transactions/confirmed/{transactionId}":
    get:
      tags:
        - Transaction routes
      summary: Get confirmed transaction information
      description: Returns confirmed transaction information given a transactionId or hash.
      operationId: getConfirmedTransaction
      parameters:
        - $ref: "#/components/parameters/transactionIdPath"
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionInfoDTO"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
        "409":
          $ref: "#/components/responses/InvalidArgument"
  /transactions/unconfirmed:
    get:
      tags:
        - Transaction routes
      summary: Search unconfirmed transactions
      description: Returns an array of unconfirmed transactions.
      operationId: searchUnconfirmedTransactions
      parameters:
        - $ref: "#/components/parameters/addressQuery"
        - $ref: "#/components/parameters/recipientAddressQuery"
        - $ref: "#/components/parameters/signerPublicKeyQuery"
        - $ref: "#/components/parameters/heightQuery"
        - $ref: "#/components/parameters/fromHeightQuery"
        - $ref: "#/components/parameters/toHeightQuery"
        - $ref: "#/components/parameters/fromTransferAmountQuery"
        - $ref: "#/components/parameters/toTransferAmountQuery"
        - $ref: "#/components/parameters/transactionTypeQuery"
        - $ref: "#/components/parameters/embedded"
        - $ref: "#/components/parameters/transferMosaicIdQuery"
        - $ref: "#/components/parameters/pageSizeQuery"
        - $ref: "#/components/parameters/pageNumberQuery"
        - $ref: "#/components/parameters/offsetQuery"
        - $ref: "#/components/parameters/orderQuery"
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionPage"
        "409":
          $ref: "#/components/responses/InvalidArgument"
    post:
      tags:
        - Transaction routes
      summary: Get unconfirmed trasactions information
      description: Returns unconfirmed transactions information for a given array of transactionIds.
      operationId: getUnconfirmedTransactions
      requestBody:
        $ref: "#/components/requestBodies/transactionIds"
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                type: array
                description: Array of transactions information.
                items:
                  $ref: "#/components/schemas/TransactionInfoDTO"
        "400":
          $ref: "#/components/responses/InvalidContent"
        "409":
          $ref: "#/components/responses/InvalidArgument"
...
servers:
  - url: "https://symbolnode.blockchain-authn.app:3001/"
    description: MAINNET
components:
  parameters:
...
    accountIdPath:
      name: accountId
      in: path
      description: Account public key or address encoded using a 32-character set.
      required: true
      schema:
        type: string
...
    transactionIdPath:
      name: transactionId
      in: path
      description: Transaction id or hash.
      required: true
      schema:
        type: string
    addressQuery:
      name: address
      in: query
      description: |
        Filter by address involved in the transaction.
        An account's address is considered to be involved in the transaction when the account is the sender, recipient, or it is required to cosign the transaction.
        This filter cannot be combined with ``recipientAddress`` and ``signerPublicKey`` query params.
      schema:
        $ref: "#/components/schemas/Address"
...
    embedded:
      name: embedded
      in: query
      description: |
        When true, the endpoint also returns all the embedded aggregate transactions.
        Otherwise, only top-level transactions used to calculate the block transactionsHash are returned.
        **Note:** This field does not work when combined with the ``address`` parameter. This is,
        embedded transactions containing the address specified through the ``address`` parameter
        will not be returned even when used with ``embedded=true``. There is no problem when using
        other parameters like ``recipientAddress`` instead.
      schema:
        type: boolean
        default: false
...
    fromHeightQuery:
      name: fromHeight
      in: query
      description: Only blocks with height greater or equal than this one are returned.
      schema:
        $ref: "#/components/schemas/Height"
    fromTransferAmountQuery:
      name: fromTransferAmount
      in: query
      description: |
        Requires providing the `transferMosaicId` filter.
        Only transfer transactions with a transfer amount of the provided mosaic id, greater or equal than this amount are returned.
      schema:
        $ref: "#/components/schemas/Amount"
    heightQuery:
      name: height
      in: query
      description: Filter by block height.
      schema:
        $ref: "#/components/schemas/Height"
...
    offsetQuery:
      name: offset
      in: query
      description: |
        Entry id at which to start pagination.
        If the ordering parameter is set to -id, the elements returned precede the identifier.
        Otherwise, newer elements with respect to the id are returned.
      schema:
        type: string
    orderQuery:
      name: order
      in: query
      description: |
        Sort responses in ascending or descending order based on the collection property set on the param ``orderBy``.
        If the request does not specify ``orderBy``, REST returns the collection ordered by id.
      schema:
        $ref: "#/components/schemas/Order"
...
    pageNumberQuery:
      name: pageNumber
      in: query
      description: Filter by page number.
      schema:
        type: integer
        minimum: 1
        default: 1
    pageSizeQuery:
      name: pageSize
      in: query
      description: Select the number of entries to return.
      schema:
        type: integer
        minimum: 10
        maximum: 100
        default: 10
...
    recipientAddressQuery:
      name: recipientAddress
      in: query
      description: Filter by address of the account receiving the transaction.
      schema:
        $ref: "#/components/schemas/Address"
...
    signerPublicKeyQuery:
      name: signerPublicKey
      in: query
      description: Filter by public key of the account signing the entity.
      schema:
        $ref: "#/components/schemas/PublicKey"
...
    toHeightQuery:
      name: toHeight
      in: query
      description: Only blocks with height smaller or equal than this one are returned.
      schema:
        $ref: "#/components/schemas/Height"
    toTransferAmountQuery:
      name: toTransferAmount
      in: query
      description: |
        Requires providing the `transferMosaicId` filter.
        Only transfer transactions with a transfer amount of the provided mosaic id, lesser or equal than this amount are returned.
      schema:
        $ref: "#/components/schemas/Amount"
    transactionTypeQuery:
      name: type
      in: query
      description: |
        Filter by transaction type.
        To filter by multiple transaction types, add more filter query params like: ``type=16974&type=16718``.
      style: form
      explode: true
      schema:
        type: array
        items:
          $ref: "#/components/schemas/TransactionTypeEnum"
    transferMosaicIdQuery:
      name: transferMosaicId
      in: query
      description: Filters transactions involving a specific ``mosaicId``.
      schema:
        $ref: "#/components/schemas/MosaicId"
  requestBodies:
...
    transactionIds:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/transactionIds"
      required: true
    transactionPayload:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/transactionPayload"
      required: true
  schemas:
...
    Height:
      type: string
      description: Height of the blockchain.
      example: 1
...
    ModelError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
...
    AccountDTO:
      type: object
      required:
        - version
        - address
        - addressHeight
        - publicKey
        - publicKeyHeight
        - accountType
        - supplementalPublicKeys
        - activityBuckets
        - mosaics
        - importance
        - importanceHeight
      properties:
        version:
          $ref: "#/components/schemas/StateVersion"
        address:
          $ref: "#/components/schemas/Address"
        addressHeight:
          $ref: "#/components/schemas/Height"
        publicKey:
          $ref: "#/components/schemas/PublicKey"
        publicKeyHeight:
          $ref: "#/components/schemas/Height"
        accountType:
          $ref: "#/components/schemas/AccountTypeEnum"
        supplementalPublicKeys:
          $ref: "#/components/schemas/SupplementalPublicKeysDTO"
        activityBuckets:
          type: array
          items:
            $ref: "#/components/schemas/ActivityBucketDTO"
        mosaics:
          type: array
          description: Mosaic units owned.
          items:
            $ref: "#/components/schemas/Mosaic"
        importance:
          $ref: "#/components/schemas/Importance"
        importanceHeight:
          $ref: "#/components/schemas/Height"
    AccountInfoDTO:
      type: object
      required:
        - id
        - account
      properties:
        id:
          type: string
          description: Internal resource identifier.
        account:
          $ref: "#/components/schemas/AccountDTO"
...
    AnnounceTransactionInfoDTO:
      type: object
      required:
        - message
      properties:
        message:
          type: string
...
    TransactionPage:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          description: Array of transactions.
          items:
            $ref: "#/components/schemas/TransactionInfoDTO"
        pagination:
          $ref: "#/components/schemas/Pagination"
...
    TransactionInfoDTO:
      type: object
      required:
        - id
        - meta
        - transaction
      properties:
        id:
          type: string
          description: Internal resource identifier.
        meta:
          anyOf:
            - $ref: "#/components/schemas/TransactionMetaDTO"
            - $ref: "#/components/schemas/EmbeddedTransactionMetaDTO"
        transaction:
          anyOf:
            - $ref: "#/components/schemas/AccountKeyLinkTransactionDTO"
            - $ref: "#/components/schemas/EmbeddedAccountKeyLinkTransactionDTO"
            - $ref: "#/components/schemas/NodeKeyLinkTransactionDTO"
            - $ref: "#/components/schemas/EmbeddedNodeKeyLinkTransactionDTO"
            - $ref: "#/components/schemas/VrfKeyLinkTransactionDTO"
            - $ref: "#/components/schemas/EmbeddedVrfKeyLinkTransactionDTO"
            - $ref: "#/components/schemas/VotingKeyLinkTransactionDTO"
            - $ref: "#/components/schemas/EmbeddedVotingKeyLinkTransactionDTO"
            - $ref: "#/components/schemas/AggregateTransactionDTO"
            - $ref: "#/components/schemas/AggregateTransactionExtendedDTO"
            - $ref: "#/components/schemas/HashLockTransactionDTO"
            - $ref: "#/components/schemas/EmbeddedHashLockTransactionDTO"
            - $ref: "#/components/schemas/SecretLockTransactionDTO"
            - $ref: "#/components/schemas/EmbeddedSecretLockTransactionDTO"
            - $ref: "#/components/schemas/SecretProofTransactionDTO"
            - $ref: "#/components/schemas/EmbeddedSecretProofTransactionDTO"
            - $ref: "#/components/schemas/AccountMetadataTransactionDTO"
            - $ref: "#/components/schemas/EmbeddedAccountMetadataTransactionDTO"
            - $ref: "#/components/schemas/MosaicMetadataTransactionDTO"
            - $ref: "#/components/schemas/EmbeddedMosaicMetadataTransactionDTO"
            - $ref: "#/components/schemas/NamespaceMetadataTransactionDTO"
            - $ref: "#/components/schemas/EmbeddedNamespaceMetadataTransactionDTO"
            - $ref: "#/components/schemas/MosaicDefinitionTransactionDTO"
            - $ref: "#/components/schemas/EmbeddedMosaicDefinitionTransactionDTO"
            - $ref: "#/components/schemas/MosaicSupplyChangeTransactionDTO"
            - $ref: "#/components/schemas/EmbeddedMosaicSupplyChangeTransactionDTO"
            - $ref: "#/components/schemas/MosaicSupplyRevocationTransactionDTO"
            - $ref: "#/components/schemas/EmbeddedMosaicSupplyRevocationTransactionDTO"
            - $ref: "#/components/schemas/NamespaceRegistrationTransactionDTO"
            - $ref: "#/components/schemas/EmbeddedNamespaceRegistrationTransactionDTO"
            - $ref: "#/components/schemas/AddressAliasTransactionDTO"
            - $ref: "#/components/schemas/EmbeddedAddressAliasTransactionDTO"
            - $ref: "#/components/schemas/MosaicAliasTransactionDTO"
            - $ref: "#/components/schemas/EmbeddedMosaicAliasTransactionDTO"
            - $ref: "#/components/schemas/MultisigAccountModificationTransactionDTO"
            - $ref: "#/components/schemas/EmbeddedMultisigAccountModificationTransactionDTO"
            - $ref: "#/components/schemas/AccountAddressRestrictionTransactionDTO"
            - $ref: "#/components/schemas/EmbeddedAccountAddressRestrictionTransactionDTO"
            - $ref: "#/components/schemas/AccountMosaicRestrictionTransactionDTO"
            - $ref: "#/components/schemas/EmbeddedAccountMosaicRestrictionTransactionDTO"
            - $ref: "#/components/schemas/AccountOperationRestrictionTransactionDTO"
            - $ref: "#/components/schemas/EmbeddedAccountOperationRestrictionTransactionDTO"
            - $ref: "#/components/schemas/MosaicGlobalRestrictionTransactionDTO"
            - $ref: "#/components/schemas/EmbeddedMosaicGlobalRestrictionTransactionDTO"
            - $ref: "#/components/schemas/MosaicAddressRestrictionTransactionDTO"
            - $ref: "#/components/schemas/EmbeddedMosaicAddressRestrictionTransactionDTO"
            - $ref: "#/components/schemas/TransferTransactionDTO"
            - $ref: "#/components/schemas/EmbeddedTransferTransactionDTO"
...
    transactionPayload:
      type: object
      properties:
        payload:
          type: string
          description: Transaction payload in hexadecimal format.
          example: 00112233445566778899AABBCCDDEEFF
  responses:
    ResourceNotFound:
      description: ResourceNotFound
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ModelError"
    InvalidContent:
      description: InvalidContent
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ModelError"
    InvalidArgument:
      description: InvalidArgument
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ModelError"